Refactor page attribute utils
