+++ src/utils/nodes/page/setPageAttributes.ts
@@ -6,10 +6,10 @@
 
 import { Node as PMNode } from "@tiptap/pm/model";
 import { Transaction } from "@tiptap/pm/state";
-import { isPageNode } from "./nodes/page/page";
-import { isBodyNode } from "./nodes/body";
-import { getPageRegionNodeAndPos } from "./pageRegion/getAttributes";
-import { setNodesTypeAttribute } from "./attributes/setAttributes";
+import { isPageNode } from "./page";
+import { isBodyNode } from "../body";
+import { getPageRegionNodeAndPos } from "../../pageRegion/getAttributes";
+import { setNodeAttribute, setNodesTypeAttribute } from "../../attributes/setAttributes";
 
 /**
  * Set a page node attribute to the given value for all page nodes in the document.
@@ -20,25 +20,6 @@ import { setNodesTypeAttribute } from "./attributes/setAttributes";
  */
 export const setPageNodesAttribute = (tr: Transaction, attr: string, value: any): boolean => {
     return setNodesTypeAttribute(tr, attr, value, setPageNodeAttribute);
-};
-
-/**
- * Set a node attribute to the given value.
- * @param tr - The transaction to apply the change to.
- * @param pos - The position of the node to set the attribute for.
- * @param node - The node to set the attribute for.
- * @param attr - The attribute to set.
- * @param value - The value to set the attribute to.
- * @returns {boolean} True if the attribute was changed, false otherwise.
- */
-const setNodeAttribute = (tr: Transaction, pos: number, node: PMNode, attr: string, value: any): boolean => {
-    const nodeAttr = node.attrs[attr];
-    const isDifferent = nodeAttr !== value;
-    if (isDifferent) {
-        tr.setNodeAttribute(pos, attr, value);
-    }
-
-    return isDifferent;
 };
 
 /**
